[MSSQL] Fix connection manager domain usage
API for the Tedious driver requires that an entry for `domain`, if present,
be at the root level of the config option passed to the constructor. This
was added in commit 91fc76be10b12e0e5a2c4a9299e5bc05a2956d58.

However, that commit deleted dialectOptions.domain from the config object
passed into the Sequelize constructor. Since most config objects are singletons,
this mutation of the config object breaks any subsequent calls to the instance
connectionManager.$connect.

This commit simply removes the call to delete the domain entry.
Fixes #4918